+++ tests/Linna/Authentication/AuthenticationTest.php
@@ -74,7 +74,7 @@ class AuthenticationTest extends TestCase
     {
         self::$session->start();
 
-        $sessionId = self::$session->id;
+        $sessionId = self::$session->getSessionId();
 
         //attemp first login
         $this->assertTrue(self::$authentication->login('root', 'password', 'root', self::$password->hash('password'), 1));
@@ -84,7 +84,7 @@ class AuthenticationTest extends TestCase
         $this->assertTrue(self::$authentication->isLogged());
 
         //check session id regeneration
-        $this->assertNotSame($sessionId, self::$session->id);
+        $this->assertNotSame($sessionId, self::$session->getSessionId());
 
         //simulate expired login
         self::$session->loginTime = \time() - 3600;
@@ -106,7 +106,7 @@ class AuthenticationTest extends TestCase
     {
         self::$session->start();
 
-        $sessionId = self::$session->id;
+        $sessionId = self::$session->getSessionId();
 
         //attemp first login
         $this->assertTrue(self::$authentication->login('root', 'password', 'root', self::$password->hash('password'), 1));
@@ -116,7 +116,7 @@ class AuthenticationTest extends TestCase
         $this->assertTrue(self::$authentication->isLogged());
 
         //check session id regeneration
-        $this->assertNotSame($sessionId, self::$session->id);
+        $this->assertNotSame($sessionId, self::$session->getSessionId());
 
         //check login data
         $this->assertTrue(self::$authentication->getLoginData()['login']);
@@ -150,7 +150,7 @@ class AuthenticationTest extends TestCase
         //attemp check if logged
         $this->assertTrue(self::$authentication->isLogged());
 
-        $sessionId = self::$session->id;
+        $sessionId = self::$session->getSessionId();
 
         //do logout
         $this->assertTrue(self::$authentication->logout());
@@ -159,7 +159,7 @@ class AuthenticationTest extends TestCase
         $this->assertFalse(self::$authentication->isLogged());
 
         //check session id regeneration
-        $this->assertNotSame($sessionId, self::$session->id);
+        $this->assertNotSame($sessionId, self::$session->getSessionId());
 
         //create new login instance
         $this->assertTrue((new Authentication(self::$session, self::$password))->isNotLogged());