+++ tests/Linna/Session/PgsqlPdoSessionHandlerTest.php
@@ -53,7 +53,7 @@ class PgsqlPdoSessionHandlerTest extends TestCase
         $pdo = (new StorageFactory('pdo', $options))->get();
 
         self::$handler = new PgsqlPdoSessionHandler($pdo);
-        self::$session = new Session(['expire' => 1800]);
+        self::$session = new Session(expire: 1800);
         self::$pdo = $pdo;
     }
 
@@ -71,11 +71,11 @@ class PgsqlPdoSessionHandlerTest extends TestCase
         $session->start();
         $session['fooData'] = 'fooData';
 
-        $this->assertEquals(2, $session->status);
-        $this->assertEquals(\session_id(), $session->id);
+        $this->assertEquals(2, $session->getStatus());
+        $this->assertEquals(\session_id(), $session->getSessionId());
         $this->assertEquals('fooData', $session['fooData']);
 
-        $oldSessionId = $session->id;
+        $oldSessionId = $session->getSessionId();
 
         $session->commit();
 
@@ -89,8 +89,8 @@ class PgsqlPdoSessionHandlerTest extends TestCase
 
         $session->start();
 
-        $this->assertEquals(2, $session->status);
-        $this->assertEquals($oldSessionId, $session->id);
+        $this->assertEquals(2, $session->getStatus());
+        $this->assertEquals($oldSessionId, $session->getSessionId());
 
         $session->destroy();
 
@@ -104,8 +104,8 @@ class PgsqlPdoSessionHandlerTest extends TestCase
 
         $this->assertEquals(0, $pdos->rowCount());
 
-        $this->assertEquals(1, $session->status);
-        $this->assertEquals('', $session->id);
+        $this->assertEquals(1, $session->getStatus());
+        $this->assertEquals('', $session->getSessionId());
         $this->assertNull($session['fooData']);
     }